Machine Learning Pipeline for - Swiggy Mobile App Food delivery platform in India
This Data set contains network of restaurants available in the 'Swiggy Mobile app' platform encompassing 150K restaurants across 600 cities in India. This pipleline explores if certain factors are the potential reasons for a restaurant to receive massive number of ratings and the same is predicted using three algorithm models with evaluation metric. Also,distribution of ratings accross all the restaurants are explored and captured to identify if below variables appeals and influences a significant number of ratings for a restaurant.